refactor: simplified auth code
