+++ console/src/pages/Dataset/DatasetVersionListCard.tsx
@@ -13,8 +13,7 @@ import { useParams } from 'react-router-dom'
 import { useFetchDatasetVersions } from '@dataset/hooks/useFetchDatasetVersions'
 import { toaster } from 'baseui/toast'
 import { ButtonLink, TextLink } from '@/components/Link'
-import WithAuth from '@/api/WithAuth'
-import { useProjectRole } from '@/domain/project/hooks/useProjectRole'
+import { WithCurrentAuth } from '@/api/WithAuth'
 
 export default function DatasetVersionListCard() {
     const [page] = usePage()
@@ -40,7 +39,6 @@ export default function DatasetVersionListCard() {
         },
         [datasetVersionsInfo, projectId, datasetId, t]
     )
-    const { role } = useProjectRole()
 
     return (
         <>
@@ -60,7 +58,7 @@ export default function DatasetVersionListCard() {
                                 datasetVersion.createdTime && formatTimestampDateTime(datasetVersion.createdTime),
                                 datasetVersion.owner && <User user={datasetVersion.owner} />,
                                 i ? (
-                                    <WithAuth role={role} id='dataset.version.revert'>
+                                    <WithCurrentAuth id='dataset.version.revert'>
                                         <ButtonLink
                                             key={datasetVersion.id}
                                             onClick={() => {
@@ -69,7 +67,7 @@ export default function DatasetVersionListCard() {
                                         >
                                             {t('Revert')}
                                         </ButtonLink>
-                                    </WithAuth>
+                                    </WithCurrentAuth>
                                 ) : (
                                     ''
                                 ),
