+++ console/src/pages/Evaluation/EvaluationListCard.tsx
@@ -25,8 +25,7 @@ import { tableNameOfSummary } from '@/domain/datastore/utils'
 import { useProject } from '@/domain/project/hooks/useProject'
 import { TextLink } from '@/components/Link'
 import { parseDecimal } from '@/utils'
-import WithAuth from '@/api/WithAuth'
-import { useProjectRole } from '@/domain/project/hooks/useProjectRole'
+import { WithCurrentAuth } from '@/api/WithAuth'
 import EvaluationListCompare from './EvaluationListCompare'
 
 const gridLayout = [
@@ -57,7 +56,6 @@ export default function EvaluationListCard() {
         },
         [evaluationsInfo, projectId]
     )
-    const { role } = useProjectRole()
 
     const store = useEvaluationStore()
 
@@ -362,7 +360,7 @@ export default function EvaluationListCard() {
                     marginBottom: 0,
                 }}
                 extra={
-                    <WithAuth role={role} id='evaluation.create'>
+                    <WithCurrentAuth id='evaluation.create'>
                         <Button
                             startEnhancer={<IconFont type='add' kind='white' />}
                             size={ButtonSize.compact}
@@ -373,7 +371,7 @@ export default function EvaluationListCard() {
                         >
                             {t('create')}
                         </Button>
-                    </WithAuth>
+                    </WithCurrentAuth>
                 }
             >
                 <StoreProvider initState={{}}>
